requirements.txt
streamlit
tensorflow-cpu
torch
torchvision
pillow
numpy
scikit-learn